pyproject.toml
[project]
name = "livekitpractice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.27.1",
    "google-genai>=1.67.0",
    "numpy>=2.4.4",
    "pyaudio>=0.2.14",
    "pyside6>=6.10.2",
    "sounddevice>=0.5.5",
]
